RT-7.5: 1. Added peer-groups & required Nokia deviations. 2. SRL does not support set-ext-community/{config,state}/method leaf. 3. It is expected in 24.10 Beta & 25.3 GA feature 9752.2
"This code is a Contribution to the OpenConfig Feature Profiles project ("Work") made under the Google Software Grant and Corporate Contributor License Agreement ("CLA") and governed by the Apache License 2.0. No other rights or licenses in or to any of Nokia's intellectual property are granted for any other purpose. This code is provided on an "as is" basis without any warranties of any kind."